journal: seperate worker mods into files
